+++ packages/camera/camera/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.8.1
+* Complete rewrite of Android plugin to fix all capture, focus, flash, and exposure issues.
+* Fixed crash when opening front-facing cameras on some legacy android devices like Sony XZ.
+* Android Flash mode works with full precapture sequence.
+* Added support for NV21 image stream format on Android.
+
 ## 0.8.0-nullsafety.3
 
 * Updates the example code listed in the [README.md](README.md), so it runs without errors when you simply copy/ paste it into a Flutter App.