+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2.0.1
+- Added support for NV21 image stream format in Android.
+
 ## 2.0.0
 
 - Stable null safety release.